+++ lib/src/ui/cms_edit_screen.dart
@@ -56,12 +56,9 @@ class CmsEditScreen<T, E> extends StatelessWidget {
                   noDataLabel: 'No fields',
                 ),
               ),
+              const ImpaktfullUiDivider(),
               Padding(
-                padding: const EdgeInsetsDirectional.only(
-                  start: 16,
-                  end: 16,
-                  bottom: 16,
-                ),
+                padding: const EdgeInsets.all(16),
                 child: ImpaktfullUiButton(
                   type: ImpaktfullUiButtonType.primary,
                   fullWidth: true,